feat: switch episode-grid component to use useCachedPromise hook
